+++ tests/creator/test_array_creator.py
@@ -153,3 +153,85 @@ def test_to_schema_no_attrs_error():
     creator = ArrayCreator(registry, "array", ("row", "col"))
     with pytest.raises(ValueError):
         creator.to_schema()
+
+
+def test_inject_dim_creator_front():
+    """Tests injecting a position into the front of the domain."""
+    registry = DataspaceRegistry()
+    SharedDim(registry, "x1", (0, 7), np.uint32)
+    SharedDim(registry, "x2", (0, 7), np.uint32)
+    SharedDim(registry, "x0", (0, 4), np.uint32)
+    creator = ArrayCreator(registry, "array", ("x1", "x2"))
+    creator.domain_creator.inject_dim_creator("x0", 0)
+    dim_names = tuple(dim_creator.name for dim_creator in creator.domain_creator)
+    assert dim_names == ("x0", "x1", "x2")
+
+
+def test_inject_dim_creator_back():
+    """Tests injecting a position into the front of the domain."""
+    registry = DataspaceRegistry()
+    SharedDim(registry, "x1", (0, 7), np.uint32)
+    SharedDim(registry, "x2", (0, 7), np.uint32)
+    SharedDim(registry, "x3", (0, 4), np.uint32)
+    creator = ArrayCreator(registry, "array", ("x1", "x2"))
+    creator.domain_creator.inject_dim_creator("x3", -1)
+    dim_names = tuple(dim_creator.name for dim_creator in creator.domain_creator)
+    assert dim_names == ("x1", "x2", "x3")
+
+
+def test_inject_dim_creator_middle():
+    """Tests injecting a position into the front of the domain."""
+    registry = DataspaceRegistry()
+    SharedDim(registry, "x0", (0, 7), np.uint32)
+    SharedDim(registry, "x2", (0, 7), np.uint32)
+    SharedDim(registry, "x1", (0, 4), np.uint32)
+    creator = ArrayCreator(registry, "array", ("x0", "x2"))
+    creator.domain_creator.inject_dim_creator("x1", 1)
+    dim_names = tuple(dim_creator.name for dim_creator in creator.domain_creator)
+    assert dim_names == ("x0", "x1", "x2")
+
+
+def test_inject_dim_attr_name_conflict_error():
+    """Tests error when injecting a dimension with name matching a current attribute
+    name."""
+    registry = DataspaceRegistry()
+    SharedDim(registry, "x0", (0, 7), np.uint32)
+    SharedDim(registry, "x2", (0, 7), np.uint32)
+    SharedDim(registry, "x1", (0, 4), np.uint32)
+    creator = ArrayCreator(registry, "array", ("x0", "x1"))
+    creator.add_attr_creator("x2", dtype=np.int32)
+    with pytest.raises(ValueError):
+        creator.domain_creator.inject_dim_creator("x2", 0)
+
+
+def test_inject_dim_name_conflict_error():
+    """Tests error when injecting a dimension with name matching a current dimension
+    name."""
+    registry = DataspaceRegistry()
+    SharedDim(registry, "x0", (0, 7), np.uint32)
+    SharedDim(registry, "x1", (0, 4), np.uint32)
+    creator = ArrayCreator(registry, "array", ("x0", "x1"))
+    with pytest.raises(ValueError):
+        creator.domain_creator.inject_dim_creator("x1", 0)
+
+
+def test_inject_dim_neg_out_of_bound_error():
+    """Tests injecting a position with a negative out-of-bound error."""
+    registry = DataspaceRegistry()
+    SharedDim(registry, "x0", (0, 7), np.uint32)
+    SharedDim(registry, "x2", (0, 7), np.uint32)
+    SharedDim(registry, "x1", (0, 4), np.uint32)
+    creator = ArrayCreator(registry, "array", ("x0", "x2"))
+    with pytest.raises(IndexError):
+        creator.domain_creator.inject_dim_creator("x1", -4)
+
+
+def test_inject_dim_pos_out_of_bound_error():
+    """Tests injecting a position with a positive out-of-bound error."""
+    registry = DataspaceRegistry()
+    SharedDim(registry, "x0", (0, 7), np.uint32)
+    SharedDim(registry, "x2", (0, 7), np.uint32)
+    SharedDim(registry, "x1", (0, 4), np.uint32)
+    creator = ArrayCreator(registry, "array", ("x0", "x2"))
+    with pytest.raises(IndexError):
+        creator.domain_creator.inject_dim_creator("x1", 3)