+++ tiledb/cf/creator.py
@@ -911,7 +911,7 @@ class ArrayCreator:
         self, dataspace_registry: DataspaceRegistry, name: str, dims: Sequence[str]
     ):
         array_registry = ArrayRegistry(dataspace_registry, name, dims)
-        return array_registry, DomainCreator(array_registry)
+        return array_registry, DomainCreator(array_registry, dataspace_registry)
 
     def attr_creator(self, key: Union[int, str]) -> AttrCreator:
         """Returns the requested attribute creator
@@ -1140,6 +1140,35 @@ class ArrayRegistry:
                 return dim_creator
         raise KeyError(f"Dimension creator with name '{key}' not found.")
 
+    def inject_dim_creator(self, dim_creator: DimCreator, position: int):
+        """Add an additional dimension into the domain of the array.
+
+        Parameters:
+            dim_creator: The dimension creator to add.
+            position: Position of the shared dimension. Negative values count backwards
+                from the end of the new number of dimensions.
+        """
+        if dim_creator.name in {dim_creator.name for dim_creator in self._dim_creators}:
+            raise ValueError(
+                f"Cannot add dimension creator `{dim_creator.name}` to this array. "
+                f"That dimension is already in use."
+            )
+        if dim_creator.name in self._attr_creators:
+            raise ValueError(
+                f"Cannot add dimension creator `{dim_creator.name}` to this array. An "
+                f"attribute creator with that name already exists."
+            )
+        index = self.ndim + 1 + position if position < 0 else position
+        if index < 0 or index > self.ndim:
+            raise IndexError(
+                f"Cannot add dimension to position {position} for an array with "
+                f"{self.ndim} dimensions."
+            )
+        print(f"INDEX: {index}")
+        self._dim_creators = (
+            self._dim_creators[:index] + (dim_creator,) + self._dim_creators[index:]
+        )
+
     @property
     def name(self) -> str:
         """Name of the array."""
@@ -1254,14 +1283,38 @@ class AttrCreator(metaclass=ABCMeta):
 class DomainCreator:
     """Creator for a TileDB domain."""
 
-    def __init__(self, array_registry):
+    def __init__(self, array_registry, dataspace_registry):
         self._array_registry = array_registry
+        self._dataspace_registry = dataspace_registry
 
     def __iter__(self):
         return self._array_registry.dim_creators()
 
     def __len__(self):
         return self.ndim
+
+    def inject_dim_creator(
+        self,
+        dim_name: str,
+        position: int,
+        tiles: Optional[Union[int, float]] = None,
+        filters: Optional[Union[tiledb.FilterList]] = None,
+    ):
+        """Adds a new dimension creator at a specified location.
+
+        Parameters:
+            dim_name: Name of the shared dimension to add to the array's domain.
+            position: Position of the shared dimension. Negative values count backwards
+                from the end of the new number of dimensions.
+            tiles: The size size for the dimension.
+            filters: Compression filters for the dimension.
+        """
+        self._array_registry.inject_dim_creator(
+            DimCreator(
+                self._dataspace_registry.get_shared_dim(dim_name), tiles, filters
+            ),
+            position,
+        )
 
     @property
     def ndim(self):